+++ tortoise/indexes.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
 class Index:
     INDEX_TYPE = ""
     INDEX_CREATE_TEMPLATE = (
-        "CREATE{index_type}INDEX {exists}{index_name} ON {table_name} ({fields}){extra};"
+        "CREATE{index_type}INDEX {index_name} ON {table_name} ({fields}){extra};"
     )
 
     def __init__(
@@ -38,10 +38,9 @@ class Index:
         self.expressions = expressions
         self.extra = ""
 
-    def get_sql(self, schema_generator: "BaseSchemaGenerator", model: "Type[Model]", safe: bool):
+    def get_sql(self, schema_generator: "BaseSchemaGenerator", model: "Type[Model]"):
         if self.fields:
             return self.INDEX_CREATE_TEMPLATE.format(
-                exists="IF NOT EXISTS " if safe else "",
                 index_name=schema_generator.quote(
                     self.name or schema_generator._generate_index_name("idx", model, self.fields)
                 ),
@@ -53,7 +52,6 @@ class Index:
 
         expressions = [f"({expression.get_sql()})" for expression in self.expressions]
         return self.INDEX_CREATE_TEMPLATE.format(
-            exists="IF NOT EXISTS " if safe else "",
             index_name=self.index_name(schema_generator, model),
             index_type=f" {self.INDEX_TYPE} ",
             table_name=schema_generator.quote(model._meta.db_table),