+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ Changelog
 Added
 ^^^^^
 - Added config_class option to pydantic model genator that allows the developer to customize the generated pydantic model's `Config` class. (#1048)
+Fixed
+^^^^^
+- Fix create index sql error. (#1202)
 
 0.19.2
 ------